[MB-14375] submitMoveForApproval refactor
