+++ src/constants/MoveHistory/EventTemplates/SubmitMoveForApproval/submitMoveForApproval.test.jsx
@@ -1,21 +1,25 @@
 import { render, screen } from '@testing-library/react';
 
 import getTemplate from 'constants/MoveHistory/TemplateManager';
-import e from 'constants/MoveHistory/EventTemplates/submitMoveForApproval';
+import e from 'constants/MoveHistory/EventTemplates/SubmitMoveForApproval/submitMoveForApproval';
 
 describe('When a customer signs and submits a move request', () => {
-  const item = {
+  const historyRecord = {
     action: 'UPDATE',
     changedValues: { status: 'NEEDS SERVICE COUNSELING' },
     eventName: 'submitMoveForApproval',
     tableName: 'moves',
   };
 
-  it('displays a confirmation message in the details column', () => {
-    const result = getTemplate(item);
+  it('matches to the proper template', () => {
+    const template = getTemplate(historyRecord);
+    expect(template).toMatchObject(e);
+  });
 
-    expect(result).toMatchObject(e);
-    render(result.getDetails(item));
+  it('displays a confirmation message in the details column', () => {
+    const template = getTemplate(historyRecord);
+
+    render(template.getDetails(historyRecord));
     expect(screen.getByText('Received customer signature')).toBeInTheDocument();
   });
 });
