+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -22,7 +22,7 @@ export { default as requestShipmentCancellation } from './requestShipmentCancell
 export { default as requestShipmentDiversion } from './requestShipmentDiversion';
 export { default as requestShipmentReweigh } from './RequestShipmentReweigh/requestShipmentReweigh';
 export { default as setFinancialReviewFlag } from './setFinancialReviewFlag';
-export { default as submitMoveForApproval } from './submitMoveForApproval';
+export { default as submitMoveForApproval } from './SubmitMoveForApproval/submitMoveForApproval';
 export { default as updateAllowance } from './UpdateAllowances/updateAllowance';
 export { default as updateBillableWeight } from './updateBillableWeight';
 export { default as updateBillableWeightAsTIO } from './updateBillableWeightAsTIO';
